requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiofiles==23.2.1
Pillow==10.1.0
python-magic==0.4.27
itsdangerous==2.1.2
starlette==0.27.0